% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(df = sample_data(), country)
}
\arguments{
\item{df}{The dataframe containing the samples}

\item{country}{An country to subset the data by}
}
\value{
A list of which the first item goes to bugs
}
\description{
Clean data to be passed on to the model function
}
\details{
This function takes a dataset that has the Structure :

id: A number or character that is the identifier of the individual
sample. This is actually not used by the function, but required as
it is passed onto the modelling tool.

country: A character or factor of the county of origin that the
analysis should be compleated for. If the dataset only has one
country just supply a variable with one repeated value.

group: Must be a factor. The first level of the factor must be the
reference group. This will frequently be humans.

ST: An integer value repersenting the type identifier

ASP, GLN, GLT, GLY, PGM, TKT, UNC: Integer values representing the
type identifier for each of the 7 MLST genes.
}
\author{
Thomas Rosendal
}

