% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataformatting.R
\name{dataformatting}
\alias{dataformatting}
\title{data formatting}
\usage{
dataformatting(DATA, UM)
}
\arguments{
\item{DATA}{A data set}

\item{UM}{An integer 0, 1, 2. Choose 0 for setting 0 as 'data
sample' for unknown source. Choose 1 for taking the mean type
frequencies as 'data sample' for unknown source. Choose 2 for
taking the type frequencies drawn from the STs that were
unique to humans, as 'data sample' for unknown source.}
}
\value{
A list of objects that will be passed to the mcmc and
    initializer
}
\description{
dataformatting
}
\details{
A function that formats the input data to be submittable to the
mcmc and initialization functions
}
\author{
Jukka Ranta
}

