% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataformattingST.R
\name{dataformatting_ST}
\alias{dataformatting_ST}
\title{data_formatting}
\usage{
dataformatting_ST(DATA, UM = 2)
}
\arguments{
\item{DATA}{A data.frame with at least A "Group" and "ST" column}

\item{UM}{An integer. Choose 0 for setting 0 as 'data sample' for
unknown source.  Choose 1 for taking the mean type frequencies
as 'data sample' for unknown source.  Choose 2 for taking the
type frequencies drawn from the STs that were unique to
humans, as 'data sample' for unknown source.}
}
\value{
A list that may be submitted to the function to initialize
    the models in R or bugs
}
\description{
Data formatting for the ST based model
}
\author{
Jukka Ranta
}

