% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializemcmc.R
\name{initialize_bugs}
\alias{initialize_bugs}
\title{initialize_bugs}
\usage{
initialize_bugs(ob)
}
\arguments{
\item{ob}{the object from the data formatting function}
}
\value{
An object that can be passed to the bug mcmc engine
}
\description{
initialize_bugs
}
\details{
Packages a number of datasets from the data cleaning function to
prepare them for open bugs
}
\author{
Jukka Ranta
}

