% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializebugsST.R
\name{initialize_bugs_ST}
\alias{initialize_bugs_ST}
\title{initialize_bugs_ST}
\usage{
initialize_bugs_ST(ob)
}
\arguments{
\item{ob}{The object for data formatting}
}
\value{
A list
}
\description{
BUGS-model with multinomial(q[Z],1)-likelihoods 
for each human isolate with unknown group indicator (Z)
}
\author{
Jukka Ranta
}

