% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializemcmc.R
\name{initialize_mcmc}
\alias{initialize_mcmc}
\title{initialize_mcmc}
\usage{
initialize_mcmc(ns, nat, MCMC, Nisolates)
}
\arguments{
\item{ns}{an integer, the number of sources}

\item{nat}{A vector the the number of alleles at each loci. In the
Order: c("ASP", "GLN", "GLT", "GLY", "PGM", "TKT", "UNC")}

\item{MCMC}{the Number of iterations for the MCMC}

\item{Nisolates}{The number of human isolates}
}
\value{
a list of the initial values for the simulation
}
\description{
A method to initialize the data objects for mcmc
}
\author{
Jukka Ranta
}

