% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_history.kilde_bugsmcmc_ST}
\alias{plot_history.kilde_bugsmcmc_ST}
\title{plot_history.kilde_bugsmcmc_ST}
\usage{
\method{plot_history}{kilde_bugsmcmc_ST}(x, burnin)
}
\arguments{
\item{x}{The kilde_bugsmcmc_ST class object}

\item{burnin}{the burnin length}
}
\value{
A plot
}
\description{
plot_history.kilde_bugsmcmc_ST
}
\details{
A plotting method for the history of the bugs object
}
\author{
Thomas Rosendal
}

