% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_history.kilde_rmcmc_ST}
\alias{plot_history.kilde_rmcmc_ST}
\title{plot_history.kilde_rmcmc_ST}
\usage{
\method{plot_history}{kilde_rmcmc_ST}(x, burnin)
}
\arguments{
\item{x}{The kilde_rmcmc_ST class object}

\item{burnin}{The burning length}
}
\value{
A plot
}
\description{
History plotting function for kilde_rmcmc_ST object
}
\author{
Thomas Rosendal
}

