% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_history}
\alias{plot_history}
\title{plot_history}
\usage{
plot_history(x, burnin)
}
\arguments{
\item{x}{The object to be plotted}

\item{burnin}{the burnin length}
}
\value{
A plot
}
\description{
plot_history
}
\details{
Defining the generic function
}
\author{
Thomas Rosendal
}

