% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_history_internal}
\alias{plot_history_internal}
\title{plot_history}
\usage{
plot_history_internal(ns, phi, MCMC, sourcenames, burnin)
}
\arguments{
\item{ns}{the number of sources}

\item{phi}{The phi object from the mcmc}

\item{MCMC}{The number of iterations}

\item{sourcenames}{A character vector of the sourcenames}

\item{burnin}{the burnin length}
}
\description{
A plotting method for showing the mcmc history of the model
}
\author{
Jukka Ranta
}

