% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_modelfit.kilde_bugsmcmc_ST}
\alias{plot_modelfit.kilde_bugsmcmc_ST}
\title{plot_modelfit.kilde_bugsmcmc_ST}
\usage{
\method{plot_modelfit}{kilde_bugsmcmc_ST}(x, burnin)
}
\arguments{
\item{x}{An object of class kilde_bugsmcmc_ST}

\item{burnin}{The burnin length}
}
\value{
A plot
}
\description{
plot_modelfit.kilde_bugsmcmc_ST
}
\author{
Thomas Rosendal
}

