% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_modelfit_internal}
\alias{plot_modelfit_internal}
\title{plot_modelfit_internal}
\usage{
plot_modelfit_internal(ns, nat, MCMC, phi, sourcenames, burnin, qASP, qGLN,
  qGLT, qGLY, qPGM, qTKT, qUNC, sourcesASP, sourcesGLN, sourcesGLT, sourcesGLY,
  sourcesPGM, sourcesTKT, sourcesUNC)
}
\arguments{
\item{ns}{The number of sources}

\item{nat}{nat}

\item{MCMC}{The number of iterations}

\item{phi}{The population attribution}

\item{sourcenames}{A character vector of source names}

\item{burnin}{the burnin length}

\item{qASP}{qASP}

\item{qGLN}{qGLN}

\item{qGLT}{qGLT}

\item{qGLY}{qGLY}

\item{qPGM}{qPGM}

\item{qTKT}{qTKT}

\item{qUNC}{qUNC}

\item{sourcesASP}{sourcesASP}

\item{sourcesGLN}{sourcesGLN}

\item{sourcesGLT}{sourcesGLT}

\item{sourcesGLY}{sourcesGLY}

\item{sourcesPGM}{sourcesPGM}

\item{sourcesTKT}{sourcesTKT}

\item{sourcesUNC}{sourcesUNC}
}
\value{
A plot
}
\description{
plot_modelfit_internal: A plot of modelfit
}
\author{
Jukka Ranta
}

