% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_population_attribution.kilde_rmcmc}
\alias{plot_population_attribution.kilde_rmcmc}
\title{plot_population_attribution.kilde_rmcmc}
\usage{
\method{plot_population_attribution}{kilde_rmcmc}(x, burnin)
}
\arguments{
\item{x}{the mcmc_object of class kilde_rmcmc}

\item{burnin}{the burnin length}
}
\value{
A plot
}
\description{
A population attribution plot for the kilde_rmcmc object
}
\author{
Thomas Rosendal
}

