% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_population_attribution.kilde_rmcmc_ST}
\alias{plot_population_attribution.kilde_rmcmc_ST}
\title{plot_population_attribution.kilde_rmcmc_ST}
\usage{
\method{plot_population_attribution}{kilde_rmcmc_ST}(x, burnin)
}
\arguments{
\item{x}{An object of class kilde_rmcmc_ST}

\item{burnin}{The burnin length}
}
\value{
A plot
}
\description{
plot_population_attribution.kilde_rmcmc_ST
}
\author{
Thomas Rosendal
}

