% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_population_attribution_internal}
\alias{plot_population_attribution_internal}
\title{plot_population_attribution_internal}
\usage{
plot_population_attribution_internal(ns, phi, burnin, MCMC, sourcenames)
}
\arguments{
\item{ns}{ns}

\item{phi}{phi}

\item{burnin}{burnin}

\item{MCMC}{MCMC}

\item{sourcenames}{sourcenames}
}
\value{
A plot
}
\description{
plot_population_attribution_internal
}
\author{
Thomas Rosendal
}

