% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_population_attribution_ST_internal}
\alias{plot_population_attribution_ST_internal}
\title{plot_population_attribution_ST_internal}
\usage{
plot_population_attribution_ST_internal(ns, STu, sourcesST, MCMC, qST, phi,
  burnin, sourcenames)
}
\arguments{
\item{ns}{ns}

\item{STu}{STu}

\item{sourcesST}{sourcesST}

\item{MCMC}{MCMC}

\item{qST}{qST}

\item{phi}{phi}

\item{burnin}{burnin}

\item{sourcenames}{sourcenames}
}
\value{
A plot
}
\description{
method tor plotting the ST model from R plot_population_attribution_ST_internal
}
\author{
Thomas Rosendal
}

