% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior_mcmc}
\alias{plot_posterior_mcmc}
\title{plot_posterior_mcmc}
\usage{
plot_posterior_mcmc(mcmc_ob, burnin)
}
\arguments{
\item{mcmc_ob}{The object returned from running the mcmc}

\item{burnin}{the burnin length}
}
\description{
plot_posterior_mcmc: A second plotting method
}
\author{
Jukka Ranta
}

