% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_r_mcmc}
\alias{plot_r_mcmc}
\title{plot_r_mcmc}
\usage{
plot_r_mcmc(mcmc_ob, burnin)
}
\arguments{
\item{mcmc_ob}{The object returned from running the mcmc}

\item{burnin}{the burnin length}
}
\description{
A plotting method for the R mcmc model
}
\author{
Jukka Ranta
}

