% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sample_attribution.kilde_rmcmc_ST}
\alias{plot_sample_attribution.kilde_rmcmc_ST}
\title{plot_sample_attribution.kilde_rmcmc_ST}
\usage{
\method{plot_sample_attribution}{kilde_rmcmc_ST}(x, burnin)
}
\arguments{
\item{x}{the model output object}

\item{burnin}{the burnin length}
}
\value{
A plot
}
\description{
plot_sample_attribution.kilde_rmcmc_ST
}
\author{
Thomas Rosendal
}

