% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sample_attribution}
\alias{plot_sample_attribution}
\title{plot_sample_attribution}
\usage{
plot_sample_attribution(x, burnin)
}
\arguments{
\item{x}{the model object}

\item{burnin}{the length of the burnin}
}
\value{
A plot
}
\description{
Method for plotting sample attribution
}
\author{
Thomas Rosendal
}

