% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sample_attribution_internal}
\alias{plot_sample_attribution_internal}
\title{plot_sample_attribution_internal}
\usage{
plot_sample_attribution_internal(Z, burnin, MCMC, ns, sourcenames)
}
\arguments{
\item{Z}{Z}

\item{burnin}{the burning length}

\item{MCMC}{The number of iterations}

\item{ns}{the number of sources}

\item{sourcenames}{A character vector of the source names}
}
\value{
A plot
}
\description{
The internal function to plot the sample attribution
}
\author{
Jukka Ranta
}

