% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_new.R
\name{run_bugs}
\alias{run_bugs}
\title{run_bugs}
\usage{
run_bugs(result, MCMC, n.burnin, FULL, model = "SA_allele_model.jag",
  n.chains = 1)
}
\arguments{
\item{result}{The result from the initialize_bugs() function}

\item{MCMC}{The number of iterations in the MCMC}

\item{n.burnin}{the burnin}

\item{FULL}{Choose 1 for full Bayesian model (semi-supervised),
but Choose 0 for separated estimation (supervised) of relative
type frequencies}

\item{model}{The path to the model file}

\item{n.chains}{the number of chains}
}
\value{
A list
}
\description{
run_bugs()
}
\details{
a function that passes the formatted data to bugs
}
\author{
Thomas Rosendal
}

