% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmcmc.R
\name{runmcmc}
\alias{runmcmc}
\title{runmcmc}
\usage{
runmcmc(result, ob, MCMC, h = 0, FULL = 0)
}
\arguments{
\item{result}{An object from the initialize mcmc function}

\item{ob}{An object from the dataformatting function}

\item{MCMC}{the number of iterations (Same as the initialize mcmc
function used)}

\item{h}{A control parameter (for avoiding very small values of q
~Dirich, maybe never needed)}

\item{FULL}{Choose 1 for full Bayesian model (semi-supervised),
but Choose 0 for separated estimation (supervised) of relative
type frequencies}
}
\value{
A list of objects to be passed onto the plot function
}
\description{
A method to run the mcmc
}
\author{
Thomas Rosendal
}

