% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmcmcST.R
\name{runmcmc_ST}
\alias{runmcmc_ST}
\title{runmcmc_ST}
\usage{
runmcmc_ST(result, ob, h = 0, FULL = 0)
}
\arguments{
\item{result}{An object from the initialize_mcmc_ST function}

\item{ob}{An object from the dataformatting function}

\item{h}{A control parameter (for avoiding very small values of q
~Dirich, maybe never needed)}

\item{FULL}{Choose 1 for full Bayesian model (semi-supervised),
but Choose 0 for separated estimation (supervised) of relative
type frequencies}
}
\value{
A list of objects to be passed onto the plot function
}
\description{
A method to run the mcmc in R: ST model
}
\author{
Jukka Ranta
}

