% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_internal}
\alias{summary_internal}
\title{summary_internal}
\usage{
summary_internal(ns, humansASP, humansGLN, humansGLT, humansGLY, humansPGM,
  humansTKT, humansUNC, sourcesASP, sourcesGLN, sourcesGLT, sourcesGLY,
  sourcesPGM, sourcesTKT, sourcesUNC, qASP, qGLN, qGLT, qGLY, qPGM, qTKT, qUNC,
  phi, burnin, MCMC, nat)
}
\arguments{
\item{ns}{ns}

\item{humansASP}{humansASP}

\item{humansGLN}{humansGLN}

\item{humansGLT}{humansGLT}

\item{humansGLY}{humansGLY}

\item{humansPGM}{humansPGM}

\item{humansTKT}{humansTKT}

\item{humansUNC}{humansUNC}

\item{sourcesASP}{sourcesASP}

\item{sourcesGLN}{sourcesGLN}

\item{sourcesGLT}{sourcesGLT}

\item{sourcesGLY}{sourcesGLY}

\item{sourcesPGM}{sourcesPGM}

\item{sourcesTKT}{sourcesTKT}

\item{sourcesUNC}{sourcesUNC}

\item{qASP}{qASP}

\item{qGLN}{qGLN}

\item{qGLT}{qGLT}

\item{qGLY}{qGLY}

\item{qPGM}{qPGM}

\item{qTKT}{qTKT}

\item{qUNC}{qUNC}

\item{phi}{phi}

\item{burnin}{burnin}

\item{MCMC}{MCMC}

\item{nat}{nat}
}
\value{
A summary
}
\description{
summary_internal
}
\details{
A summary of a model run
}
\author{
Jukka Ranta
}

