% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_kilde.kilde_bugsmcmc}
\alias{summary_kilde.kilde_bugsmcmc}
\title{summary_kilde.kilde_bugsmcmc}
\usage{
\method{summary_kilde}{kilde_bugsmcmc}(x, burnin)
}
\arguments{
\item{x}{A kilde_bugsmcmc object}

\item{burnin}{the burnin length}
}
\value{
A summary
}
\description{
A summary method for the bugsmcmc function
}
\author{
Thomas Rosendal
}

