% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_kilde.kilde_rmcmc}
\alias{summary_kilde.kilde_rmcmc}
\title{summary_kilde.kilde_rmcmc}
\usage{
\method{summary_kilde}{kilde_rmcmc}(x, burnin)
}
\arguments{
\item{x}{kilde_rmcmc object}

\item{burnin}{The burnin length}
}
\value{
A summary
}
\description{
A summary method for the kilde_rmcmc object
}
\author{
Thomas Rosendal
}

