% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_kilde.kilde_rmcmc_ST}
\alias{summary_kilde.kilde_rmcmc_ST}
\title{summary_kilde.kilde_rmcmc_ST}
\usage{
\method{summary_kilde}{kilde_rmcmc_ST}(x, burnin)
}
\arguments{
\item{x}{An object of class kilde_rmcmc_ST}

\item{burnin}{the burnin length}
}
\value{
A list
}
\description{
summary_kilde.kilde_rmcmc_ST
}
\author{
Thomas Rosendal
}

