% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_kilde}
\alias{summary_kilde}
\title{summary}
\usage{
summary_kilde(x, burnin)
}
\arguments{
\item{x}{The object to be summarized}

\item{burnin}{the burnin length}
}
\value{
A list
}
\description{
summary of models
}
\details{
Defining the generic function
}
\author{
Thomas Rosendal
}

