% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_kilde_ST_internal}
\alias{summary_kilde_ST_internal}
\title{summary_kilde_ST_internal}
\usage{
summary_kilde_ST_internal(ns, phi, burnin, MCMC, humansST, STu, sourcesST, qST)
}
\arguments{
\item{ns}{ns}

\item{phi}{phi}

\item{burnin}{burnin}

\item{MCMC}{MCMC}

\item{humansST}{humansST}

\item{STu}{STu}

\item{sourcesST}{sourcesST}

\item{qST}{qST}
}
\value{
A list
}
\description{
summary_kilde_ST_internal
}
\author{
Thomas Rosendal
}

