% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(df = sample_data(), country)
}
\arguments{
\item{df}{}
}
\value{
list
}
\description{
Clean data to be passed on to the model function
}
\author{
Thomas Rosendal
}

