% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_population_attribution.kilde_bugsmcmc}
\alias{plot_population_attribution.kilde_bugsmcmc}
\title{plot_population_attribution.kilde_bugsmcmc}
\usage{
\method{plot_population_attribution}{kilde_bugsmcmc}(x, burnin)
}
\arguments{
\item{x}{the kilde_bugsmcmc object}

\item{burnin}{The burnin length}
}
\value{
A plot
}
\description{
plot_population_attribution.kilde_bugsmcmc
}
\details{
A method for ploting the population attribution from the bugs code
}
\author{
Thomas Rosendal
}

