% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_population_attribution}
\alias{plot_population_attribution}
\title{plot_population_attribution}
\usage{
plot_population_attribution(x, burnin)
}
\arguments{
\item{x}{x}

\item{burnin}{burnin}
}
\value{
A plot
}
\description{
plot_population_attribution
}
\author{
Thomas Rosendal
}

