% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(model = "SA_allelemodel1.jag", data, inits, parameters,
  n.chains = 1, n.burnin = 99, n.iter = 999)
}
\arguments{
\item{model}{The name of the model
This is the file name in the inst/models directory of the package}

\item{data}{A dataset}

\item{inits}{An inits function}

\item{parameters}{Parameters to monitor}

\item{n.chains}{The number of chains}

\item{n.burnin}{The burning length}

\item{n.iter}{The number of iterations}
}
\value{
A model object
}
\description{
A method to run a model in bugs
}
\author{
Thomas Rosendal
}

